pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "bcrypt==4.0.1",
    "fastapi[all]>=0.126.0",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "scipy>=1.17.0",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.38.0",
]
